+++ eelbrain/__init__.py
@@ -25,7 +25,7 @@ from ._mne import labels_from_clusters, morph_source_space
 from ._ndvar import (
     Butterworth, concatenate, convolve, cross_correlation, cwt_morlet, dss,
     filter_data, find_intervals, find_peaks, label_operator,
-    neighbor_correlation, psd_welch, resample, segment,
+    neighbor_correlation, psd_welch, resample, segment, set_parc,
 )
 from ._trf import boosting, BoostingResult
 from ._utils import set_log_level